requirements.txt
numpy== 1.26.4
pandas== 2.2.3
matplotlib== 3.9.2
seaborn== 0.13.2
scikit-learn== 1.5.1
torch== 2.7.1
torchvision== 0.22.1
Pillow== 10.4.0
Python==3.12.7